requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.4
filelock==3.9.0
Flask==2.2.2
huggingface-hub==0.11.1
idna==3.4
imageio==2.23.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
networkx==2.8.8
nltk==3.8
numpy==1.24.1
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.3.0
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tifffile==2022.10.10
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
